Update LIAM EGO to work with  evaluation
