+++ ego_agent_training/liam_ego.py
@@ -14,13 +14,13 @@ import optax
 import hydra
 from flax.training.train_state import TrainState
 
+from agents.initialize_agents import initialize_liam_agent
 from agents.population_interface import AgentPopulation
-from marl.liam_utils import Transition, initialize_encoder_decoder, _create_minibatches
+from marl.liam_utils import Transition
 from common.run_episodes import run_episodes
 from common.plot_utils import get_stats, get_metric_names
-from marl.ppo_utils import unbatchify
+from marl.ppo_utils import unbatchify, _create_minibatches
 from common.save_load_utils import save_train_run
-from ego_agent_training.utils import initialize_ego_agent
 from envs import make_env
 from envs.log_wrapper import LogWrapper
 from evaluation.agent_loader_from_config import initialize_rl_agent_from_config
@@ -31,7 +31,6 @@ logging.basicConfig(level=logging.INFO)
 
 def train_liam_ego_agent(config, env, train_rng, 
                         ego_policy, init_ego_params, n_ego_train_seeds,
-                        encoder, decoder, init_encoder_decoder_params,
                         partner_population: AgentPopulation,
                         partner_params
                         ):
@@ -88,8 +87,8 @@ def train_liam_ego_agent(config, env, train_rng,
                 )
 
             train_state = TrainState.create(
-                apply_fn=ego_policy.network.apply,
-                params=init_ego_params,
+                apply_fn=ego_policy.policy.network.apply,
+                params=init_ego_params['policy'],
                 tx=tx,
             )
 
@@ -99,8 +98,8 @@ def train_liam_ego_agent(config, env, train_rng,
             )
 
             encoder_decoder_train_state = TrainState.create(
-                apply_fn=encoder.model.apply,
-                params=init_encoder_decoder_params,
+                apply_fn=ego_policy.encoder.model.apply,
+                params={'encoder': init_ego_params['encoder'], 'decoder': init_ego_params['decoder']},
                 tx=encoder_decoder_tx,
             )
 
@@ -111,7 +110,7 @@ def train_liam_ego_agent(config, env, train_rng,
                 2. Step environment using sampled actions
                 3. Return state, reward, ...
                 """
-                train_state, encoder_decoder_train_state, env_state, prev_obs, prev_done, ego_hstate, ego_encoder_hstate, partner_hstate, partner_indices, rng = runner_state
+                train_state, encoder_decoder_train_state, env_state, prev_obs, prev_done, ego_hstate, partner_hstate, partner_indices, rng = runner_state
                 rng, actor_rng, partner_rng, step_rng = jax.random.split(rng, 4)
 
                  # Get available actions for agent 0 from environment state
@@ -120,18 +119,13 @@ def train_liam_ego_agent(config, env, train_rng,
                 avail_actions_0 = avail_actions["agent_0"].astype(jnp.float32)
                 avail_actions_1 = avail_actions["agent_1"].astype(jnp.float32)
 
-                # Agent 0 Encoder
-                embedded_obs_0, ego_encoder_hstate = encoder.compute_embedding(
-                    params=encoder_decoder_train_state.params['encoder'],
-                    hstate=ego_encoder_hstate,
-                    obs=prev_obs["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"], -1), 
-                    done=prev_done["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"])
-                )
                                 
                 # Agent_0 (ego) action, value, log_prob
                 act_0, val_0, pi_0, ego_hstate = ego_policy.get_action_value_policy(
-                    params=train_state.params,
-                    obs=embedded_obs_0,
+                    params={"encoder": encoder_decoder_train_state.params["encoder"], 
+                            "decoder": encoder_decoder_train_state.params["decoder"],
+                            "policy": train_state.params},
+                    obs=prev_obs["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"], -1),
                     done=prev_done["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"]),
                     avail_actions=avail_actions_0,
                     hstate=ego_hstate,
@@ -195,7 +189,7 @@ def train_liam_ego_agent(config, env, train_rng,
                     partner_action=jax.nn.one_hot(act_1, env.action_space(env.agents[1]).n)
                 )
                 new_runner_state = (train_state, encoder_decoder_train_state, env_state_next, obs_next, done_next, 
-                                    ego_hstate, ego_encoder_hstate, new_partner_hstate, updated_partner_indices, rng)
+                                    ego_hstate, new_partner_hstate, updated_partner_indices, rng)
                 return new_runner_state, transition
 
             def _calculate_gae(traj_batch, last_val):
@@ -225,34 +219,23 @@ def train_liam_ego_agent(config, env, train_rng,
 
             def _update_minbatch(init_state, batch_info):
                 train_state, encoder_decoder_train_state = init_state
-                init_ego_hstate, init_ego_encoder_hstate, traj_batch, advantages, returns = batch_info
-                def _loss_fn(params, encoder_decoder_params, init_ego_hstate, init_ego_encoder_hstate, traj_batch, gae, target_v):
+                init_ego_hstate, traj_batch, advantages, returns = batch_info
+                def _loss_fn(params, encoder_decoder_params, init_ego_hstate, traj_batch, gae, target_v):
 
-                    # Agent 0 Encoder
-                    embedded_batch_obs, _ = encoder.compute_embedding(
-                        params=encoder_decoder_params['encoder'],
-                        hstate=init_ego_encoder_hstate,
+
+                    _, value, pi, recon_loss1, recon_loss2, _ = ego_policy.evalute(
+                        params={"encoder": encoder_decoder_params["encoder"], 
+                                "decoder": encoder_decoder_params["decoder"],
+                                "policy": params},
                         obs=traj_batch.obs,
-                        done=traj_batch.done
-                    )
-
-                    _, value, pi, _ = ego_policy.get_action_value_policy(
-                        params=params, 
-                        obs=jax.lax.stop_gradient(embedded_batch_obs),
                         done=traj_batch.done,
                         avail_actions=traj_batch.avail_actions,
                         hstate=init_ego_hstate,
-                        rng=jax.random.PRNGKey(0) # only used for action sampling, which is unused here
-                    )
-                    log_prob = pi.log_prob(traj_batch.action)
-
-                    # Reconstruction Loss
-                    recon_loss1, recon_loss2 = decoder.evaluate(
-                        params=encoder_decoder_params['decoder'],
-                        embedding=embedded_batch_obs,
+                        rng=jax.random.PRNGKey(0), # only used for action sampling, which is unused here
                         modelled_agent_obs=traj_batch.partner_obs,
                         modelled_agent_act=traj_batch.partner_action
                     )
+                    log_prob = pi.log_prob(traj_batch.action)
                     recon_loss = (recon_loss1.mean() + recon_loss2.mean())
 
                     # Value loss
@@ -284,7 +267,7 @@ def train_liam_ego_agent(config, env, train_rng,
 
                 grad_fn = jax.value_and_grad(_loss_fn, argnums=(0,1), has_aux=True)
                 (loss_val, aux_vals), (grads, encoder_decoder_grads) = grad_fn(
-                    train_state.params, encoder_decoder_train_state.params, init_ego_hstate, init_ego_encoder_hstate, traj_batch, advantages, returns)
+                    train_state.params, encoder_decoder_train_state.params, init_ego_hstate, traj_batch, advantages, returns)
                 train_state = train_state.apply_gradients(grads=grads)
                 encoder_decoder_train_state = encoder_decoder_train_state.apply_gradients(grads=encoder_decoder_grads)
                 
@@ -302,13 +285,13 @@ def train_liam_ego_agent(config, env, train_rng,
                 return (train_state, encoder_decoder_train_state), (loss_val, aux_vals, avg_grad_norm, encoder_decoder_avg_grad_norm)
 
             def _update_epoch(update_state, unused):
-                train_state, encoder_decoder_train_state, init_ego_hstate, init_ego_encoder_hstate, traj_batch, advantages, targets, rng = update_state
+                train_state, encoder_decoder_train_state, init_ego_hstate, traj_batch, advantages, targets, rng = update_state
                 rng, perm_rng = jax.random.split(rng)
-                minibatches = _create_minibatches(traj_batch, advantages, targets, init_ego_hstate, init_ego_encoder_hstate, config["NUM_CONTROLLED_ACTORS"], config["NUM_MINIBATCHES"], perm_rng)
+                minibatches = _create_minibatches(traj_batch, advantages, targets, init_ego_hstate, config["NUM_CONTROLLED_ACTORS"], config["NUM_MINIBATCHES"], perm_rng)
                 (train_state, encoder_decoder_train_state), losses_and_grads = jax.lax.scan(
                     _update_minbatch, (train_state, encoder_decoder_train_state), minibatches
                 )
-                update_state = (train_state, encoder_decoder_train_state, init_ego_hstate, init_ego_encoder_hstate, traj_batch, advantages, targets, rng)
+                update_state = (train_state, encoder_decoder_train_state, init_ego_hstate, traj_batch, advantages, targets, rng)
                 return update_state, losses_and_grads
 
             def _update_step(update_runner_state, unused):
@@ -317,30 +300,24 @@ def train_liam_ego_agent(config, env, train_rng,
                 2. Compute advantage
                 3. LIAM updates
                 """
-                (train_state, encoder_decoder_train_state, last_env_state, last_obs, last_done, last_ego_hstate, last_ego_encoder_hstate, last_partner_hstate, last_partner_indices, rng, update_steps) = update_runner_state
+                (train_state, encoder_decoder_train_state, last_env_state, last_obs, last_done, last_ego_hstate, last_partner_hstate, last_partner_indices, rng, update_steps) = update_runner_state
 
                 # 1) rollout
-                runner_state = (train_state, encoder_decoder_train_state, last_env_state, last_obs, last_done, last_ego_hstate, last_ego_encoder_hstate, last_partner_hstate, last_partner_indices, rng)
+                runner_state = (train_state, encoder_decoder_train_state, last_env_state, last_obs, last_done, last_ego_hstate, last_partner_hstate, last_partner_indices, rng)
                 runner_state, traj_batch = jax.lax.scan(
                     _env_step, runner_state, None, config["ROLLOUT_LENGTH"])
-                (train_state, encoder_decoder_train_state, env_state, obs, done, ego_hstate, ego_encoder_hstate, partner_hstate, partner_indices, rng) = runner_state
+                (train_state, encoder_decoder_train_state, env_state, obs, done, ego_hstate, partner_hstate, partner_indices, rng) = runner_state
 
                 # 2) advantage
                 # Get available actions for agent 0 from environment state
                 avail_actions_0 = jax.vmap(env.get_avail_actions)(env_state.env_state)["agent_0"].astype(jnp.float32)
 
-                # Agent 0 Encoder
-                embedded_obs, _ = encoder.compute_embedding(
-                    params=encoder_decoder_train_state.params['encoder'],
-                    hstate=ego_encoder_hstate,
-                    obs=obs["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"], -1),
-                    done=done["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"])
-                )
-
                 # Get final value estimate for completed trajectory
                 _, last_val, _, _ = ego_policy.get_action_value_policy(
-                    params=train_state.params, 
-                    obs=embedded_obs,
+                    params={"encoder": encoder_decoder_train_state.params["encoder"], 
+                            "decoder": encoder_decoder_train_state.params["decoder"],
+                            "policy": train_state.params}, 
+                    obs=obs["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"], -1),
                     done=done["agent_0"].reshape(1, config["NUM_CONTROLLED_ACTORS"]),
                     avail_actions=jax.lax.stop_gradient(avail_actions_0),
                     hstate=ego_hstate,
@@ -351,12 +328,10 @@ def train_liam_ego_agent(config, env, train_rng,
 
                 # 3) LIAM update
                 init_ego_hstate = ego_policy.init_hstate(config["NUM_CONTROLLED_ACTORS"])
-                init_ego_encoder_hstate = encoder.init_hstate(batch_size=config["NUM_CONTROLLED_ACTORS"])
                 update_state = (
                     train_state,
                     encoder_decoder_train_state,
-                    init_ego_hstate, # shape is (num_controlled_actors, gru_hidden_dim) with all-0s value
-                    init_ego_encoder_hstate, # shape is (num_controlled_actors, encoder_hidden_dim) with all-0s value
+                    init_ego_hstate, # (shape is (num_controlled_actors, encoder_hidden_dim) with all-0s value, shape is (num_controlled_actors, gru_hidden_dim) with all-0s value)
                     traj_batch, # obs has shape (rollout_len, num_controlled_actors, -1)
                     advantages,
                     targets,
@@ -377,7 +352,7 @@ def train_liam_ego_agent(config, env, train_rng,
                 metric["avg_grad_norm"] = avg_grad_norm
                 metric["encoder_decoder_avg_grad_norm"] = encoder_decoder_avg_grad_norm
                 new_runner_state = (train_state, encoder_decoder_train_state, env_state, obs, done, 
-                                    ego_hstate, ego_encoder_hstate, partner_hstate, partner_indices, 
+                                    ego_hstate, partner_hstate, partner_indices, 
                                     rng, update_steps + 1)
                 return (new_runner_state, metric)
 
@@ -394,7 +369,7 @@ def train_liam_ego_agent(config, env, train_rng,
             max_episode_steps = config["ROLLOUT_LENGTH"]
             
             def _update_step_with_ckpt(state_with_ckpt, unused):
-                (update_state, checkpoint_array, encoder_decoder_checkpoint_array, ckpt_idx, init_eval_last_info) = state_with_ckpt
+                (update_state, checkpoint_array, ckpt_idx, init_eval_last_info) = state_with_ckpt
 
                 # Single LIAM update
                 new_update_state, metric = _update_step(
@@ -402,7 +377,7 @@ def train_liam_ego_agent(config, env, train_rng,
                     None
                 )
                 (train_state, encoder_decoder_train_state, env_state, obs, done, 
-                 ego_hstate, ego_encoder_hstate, partner_hstate, partner_indices, 
+                 ego_hstate, partner_hstate, partner_indices, 
                  rng, update_steps) = new_update_state
 
                 # update steps is 1-indexed because it was incremented at the end of the update step
@@ -411,40 +386,41 @@ def train_liam_ego_agent(config, env, train_rng,
 
 
                 def store_and_eval_ckpt(args):
-                    ckpt_arr, encode_decode_ckpt_arr, cidx, rng, prev_eval_ret_info = args
+                    ckpt_arr, cidx, rng, prev_eval_ret_info = args
                     new_ckpt_arr = jax.tree.map(
                         lambda c_arr, p: c_arr.at[cidx].set(p),
-                        ckpt_arr, train_state.params
-                    )
-
-                    new_encode_decode_ckpt_arr = jax.tree.map(
-                        lambda c_arr, p: c_arr.at[cidx].set(p),
-                        encode_decode_ckpt_arr, encoder_decoder_train_state.params
+                        ckpt_arr, {"encoder": encoder_decoder_train_state.params["encoder"], 
+                                   "decoder": encoder_decoder_train_state.params["decoder"],
+                                   "policy": train_state.params}
                     )
 
                     eval_partner_indices = jnp.arange(num_total_partners)
                     gathered_params = partner_population.gather_agent_params(partner_params, eval_partner_indices)
                     
                     rng, eval_rng = jax.random.split(rng)
+                    # TODO: Fix eval to use encoder
                     eval_eps_last_infos = jax.vmap(lambda x: run_episodes(
-                        eval_rng, env, agent_0_param=train_state.params, agent_0_policy=ego_policy, 
+                        eval_rng, env, agent_0_param={"encoder": encoder_decoder_train_state.params["encoder"], 
+                                                      "decoder": encoder_decoder_train_state.params["decoder"],
+                                                      "policy": train_state.params}, 
+                        agent_0_policy=ego_policy, 
                         agent_1_param=x, agent_1_policy=partner_population.policy_cls, 
                         max_episode_steps=max_episode_steps, 
                         num_eps=config["NUM_EVAL_EPISODES"]))(gathered_params)
-                    return (new_ckpt_arr, new_encode_decode_ckpt_arr, cidx + 1, rng, eval_eps_last_infos)
+                    return (new_ckpt_arr, cidx + 1, rng, eval_eps_last_infos)
                 
                 def skip_ckpt(args):
                     return args
 
-                (checkpoint_array, encoder_decoder_checkpoint_array, ckpt_idx, rng, eval_last_infos) = jax.lax.cond(
-                    to_store, store_and_eval_ckpt, skip_ckpt, (checkpoint_array, encoder_decoder_checkpoint_array, ckpt_idx, rng, init_eval_last_info)
+                (checkpoint_array, ckpt_idx, rng, eval_last_infos) = jax.lax.cond(
+                    to_store, store_and_eval_ckpt, skip_ckpt, (checkpoint_array, ckpt_idx, rng, init_eval_last_info)
                 )
 
                 metric["eval_ep_last_info"] = eval_last_infos
                 return ((train_state, encoder_decoder_train_state, env_state, obs, done, 
-                         ego_hstate, ego_encoder_hstate, partner_hstate, partner_indices, 
+                         ego_hstate, partner_hstate, partner_indices, 
                          rng, update_steps),
-                         checkpoint_array, encoder_decoder_checkpoint_array, ckpt_idx, eval_last_infos), metric
+                         checkpoint_array, ckpt_idx, eval_last_infos), metric
 
             # Init envs & partner indices
             rng, reset_rng = jax.random.split(rng)
@@ -458,11 +434,13 @@ def train_liam_ego_agent(config, env, train_rng,
             
             # Init ego and partner hstates
             init_ego_hstate = ego_policy.init_hstate(config["NUM_CONTROLLED_ACTORS"])
-            init_ego_encoder_hstate = encoder.init_hstate(batch_size=config["NUM_CONTROLLED_ACTORS"])
             init_partner_hstate = partner_population.init_hstate(config["NUM_UNCONTROLLED_ACTORS"])
 
-            checkpoint_array = init_ckpt_array(train_state.params)
-            encoder_decoder_checkpoint_array = init_ckpt_array(encoder_decoder_train_state.params)
+            checkpoint_array = init_ckpt_array(
+                {"encoder": encoder_decoder_train_state.params["encoder"], 
+                 "decoder": encoder_decoder_train_state.params["decoder"],
+                 "policy": train_state.params}
+            )
             ckpt_idx = 0
 
             rng, rng_eval, rng_train = jax.random.split(rng, 3)
@@ -471,7 +449,10 @@ def train_liam_ego_agent(config, env, train_rng,
             gathered_params = partner_population.gather_agent_params(partner_params, eval_partner_indices)
             eval_eps_last_infos = jax.vmap(lambda x: run_episodes(
                         rng_eval, env, 
-                        agent_0_param=train_state.params, agent_0_policy=ego_policy, 
+                        agent_0_param={"encoder": encoder_decoder_train_state.params["encoder"], 
+                                       "decoder": encoder_decoder_train_state.params["decoder"],
+                                       "policy": train_state.params}, 
+                        agent_0_policy=ego_policy, 
                         agent_1_param=x, agent_1_policy=partner_population.policy_cls, 
                         max_episode_steps=max_episode_steps, 
                         num_eps=config["NUM_EVAL_EPISODES"]))(gathered_params)
@@ -487,13 +468,12 @@ def train_liam_ego_agent(config, env, train_rng,
                 init_obs,
                 init_done,
                 init_ego_hstate,
-                init_ego_encoder_hstate,
                 init_partner_hstate,
                 init_partner_indices,
                 rng_train,
                 update_steps
             )
-            state_with_ckpt = (update_runner_state, checkpoint_array, encoder_decoder_checkpoint_array, ckpt_idx, eval_eps_last_infos)
+            state_with_ckpt = (update_runner_state, checkpoint_array, ckpt_idx, eval_eps_last_infos)
             
             state_with_ckpt, metrics = jax.lax.scan(
                 _update_step_with_ckpt,
@@ -501,12 +481,13 @@ def train_liam_ego_agent(config, env, train_rng,
                 xs=None,
                 length=config["NUM_UPDATES"]
             )
-            (final_runner_state, checkpoint_array, encoder_decoder_checkpoint_array, final_ckpt_idx, eval_eps_last_infos) = state_with_ckpt
+            (final_runner_state, checkpoint_array, final_ckpt_idx, eval_eps_last_infos) = state_with_ckpt
             out = {
-                "final_params": final_runner_state[0].params,
+                "final_params": {"encoder": final_runner_state[1].params["encoder"], 
+                                 "decoder": final_runner_state[1].params["decoder"],
+                                 "policy": final_runner_state[0].params},
                 "metrics": metrics,  # shape (NUM_UPDATES, ...)
                 "checkpoints": checkpoint_array,
-                "encoder_decoder_checkpoints": encoder_decoder_checkpoint_array,
             }
             return out
         return train
@@ -532,7 +513,7 @@ def run_ego_training(config, wandb_logger):
     env = LogWrapper(env)
 
     rng = jax.random.PRNGKey(algorithm_config["TRAIN_SEED"])
-    rng, init_partner_rng, init_ego_rng, init_ego_encoder_decoder_rng, train_rng = jax.random.split(rng, 5)
+    rng, init_partner_rng, init_ego_rng, train_rng = jax.random.split(rng, 4)
     
 
     partner_agent_config = dict(algorithm_config["partner_agent"])
@@ -549,10 +530,7 @@ def run_ego_training(config, wandb_logger):
     )
     
     # Initialize ego agent policy
-    ego_policy, init_ego_params = initialize_ego_agent(algorithm_config, env, init_ego_rng)
-
-    # Initialize ego agent encoder and decoder
-    encoder, decoder, init_ecoder_decoder_params = initialize_encoder_decoder(algorithm_config, env, init_ego_encoder_decoder_rng)
+    ego_policy, init_ego_params = initialize_liam_agent(algorithm_config, env, init_ego_rng)
     
     log.info("Starting ego agent training...")
     start_time = time.time()
@@ -564,9 +542,6 @@ def run_ego_training(config, wandb_logger):
         train_rng=train_rng,
         ego_policy=ego_policy,
         init_ego_params=init_ego_params,
-        encoder=encoder,
-        decoder=decoder,
-        init_encoder_decoder_params=init_ecoder_decoder_params,
         n_ego_train_seeds=algorithm_config["NUM_EGO_TRAIN_SEEDS"],
         partner_population=partner_population,
         partner_params=flattened_partner_params
@@ -578,7 +553,7 @@ def run_ego_training(config, wandb_logger):
     metric_names = get_metric_names(config["ENV_NAME"])
     log_metrics(config, out, wandb_logger, metric_names)
     
-    return out["final_params"], ego_policy, init_ego_params, encoder, decoder, init_ecoder_decoder_params
+    return out["final_params"], ego_policy, init_ego_params
 
 def log_metrics(config, train_out, logger, metric_names: tuple):
     """Process training metrics and log them using the provided logger.
